requirements.txt
--find-links=https://girder.github.io/large_image_wheels GDAL
geopandas
leafmap
owslib
streamlit
altair                       5.4.1
annotated-types              0.7.0
anywidget                    0.9.13
asttokens                    2.4.1
attrs                        24.2.0
beautifulsoup4               4.12.3
blinker                      1.8.2
bqplot                       0.12.43
branca                       0.8.0
cachetools                   5.5.0
certifi                      2024.8.30
charset-normalizer           3.4.0
click                        8.1.7
colorama                     0.4.6
colour                       0.1.5
comm                         0.2.2
contourpy                    1.3.0
cycler                       0.12.1
decorator                    5.1.1
duckdb                       1.1.2
executing                    2.1.0
filelock                     3.16.1
folium                       0.18.0
fonttools                    4.54.1
gdown                        5.2.0
geojson                      3.1.0
geopandas                    1.0.1
gitdb                        4.0.11
GitPython                    3.1.43
google-ai-generativelanguage 0.6.10
google-api-core              2.21.0
google-api-python-client     2.149.0
google-auth                  2.35.0
google-auth-httplib2         0.2.0
google-generativeai          0.8.3
googleapis-common-protos     1.65.0
grpcio                       1.67.0
grpcio-status                1.67.0
httplib2                     0.22.0
idna                         3.10
ipyevents                    2.0.2
ipyfilechooser               0.6.0
ipyleaflet                   0.19.2
ipython                      8.29.0
ipytree                      0.2.2
ipyvue                       1.11.1
ipyvuetify                   1.10.0
ipywidgets                   8.1.5
jedi                         0.19.1
Jinja2                       3.1.4
joblib                       1.4.2
jsonschema                   4.23.0
jsonschema-specifications    2024.10.1
jupyter-leaflet              0.19.2
jupyterlab_widgets           3.0.13
kiwisolver                   1.4.7
leafmap                      0.38.10
lxml                         5.3.0
markdown-it-py               3.0.0
MarkupSafe                   3.0.2
matplotlib                   3.9.2
matplotlib-inline            0.1.7
mdurl                        0.1.2
narwhals                     1.10.0
numpy                        2.1.2
OWSLib                       0.32.0
packaging                    24.1
pandas                       2.2.3
parso                        0.8.4
pillow                       10.4.0
pip                          24.2
plotly                       5.24.1
prompt_toolkit               3.0.48
proto-plus                   1.25.0
protobuf                     5.28.3
psutil                       6.1.0
psygnal                      0.11.1
pure_eval                    0.2.3
pyarrow                      17.0.0
pyasn1                       0.6.1
pyasn1_modules               0.4.1
pydantic                     2.9.2
pydantic_core                2.23.4
pydeck                       0.9.1
Pygments                     2.18.0
pyogrio                      0.10.0
pyparsing                    3.2.0
pyproj                       3.7.0
pyshp                        2.3.1
PySocks                      1.7.1
pystac                       1.11.0
pystac-client                0.8.5
python-box                   7.2.0
python-dateutil              2.9.0.post0
python-dotenv                1.0.1
pytz                         2024.2
PyYAML                       6.0.2
referencing                  0.35.1
requests                     2.32.3
rich                         13.9.3
rpds-py                      0.20.0
rsa                          4.9
scikit-learn                 1.5.2
scipy                        1.14.1
scooby                       0.10.0
setuptools                   75.2.0
shapely                      2.0.6
six                          1.16.0
smmap                        5.0.1
soupsieve                    2.6
stack-data                   0.6.3
streamlit                    1.39.0
streamlit_folium             0.23.1
tenacity                     9.0.0
threadpoolctl                3.5.0
toml                         0.10.2
tornado                      6.4.1
tqdm                         4.66.5
traitlets                    5.14.3
traittypes                   0.2.1
typing_extensions            4.12.2
tzdata                       2024.2
uritemplate                  4.1.1
urllib3                      2.2.3
watchdog                     5.0.3
wcwidth                      0.2.13
whitebox                     2.3.5
whiteboxgui                  2.3.0
widgetsnbextension           4.0.13
xgboost                      2.1.2
xyzservices                  2024.9.0